// src/components/ContextMenu/Types.ts
export enum ContextMenuItem {
  IMMERSIVE = 'immersive',
  MINI_PLAYER = 'miniPlayer',
  SHARE = 'share',
  PROPERTIES = 'properties',
  CREATE_STATION = 'createStation',
  GO_TO_ARTIST = 'goToArtist',
  GO_TO_ALBUM = 'goToAlbum',
  SHUFFLE = 'shuffle',
  REMOVE_FROM_QUEUE = 'removeFromQueue',
  ADD_TO_PLAYLIST = 'addToPlaylist',
  FAVORITE_QUICK = 'favoriteQuick',
  CREATE_STATION_QUICK = 'createStationQuick',
  LOVE_QUICK = 'loveQuick',
  DISLIKE_QUICK = 'dislikeQuick',
  ADD_TO_LIBRARY_QUICK = 'addToLibraryQuick',
  PLAY_NEXT_QUICK = 'playNextQuick',
  PLAY_LAST_QUICK = 'playLastQuick'
}